refactor session list view and history view
